factories
- added product factory
